+++ src/app/services/reducers/index.ts
@@ -5,9 +5,9 @@ import collections from '../slices/collections.slice';
 import devxApi from '../slices/devxapi.slice';
 import dimensions from '../slices/dimensions.slice';
 import graphExplorerMode from '../slices/explorer-mode.slice';
-import permissionGrants from '../slices/grants.slice';
 import graphResponse from '../slices/graph-response.slice';
 import history from '../slices/history.slice';
+import permissionGrants from '../slices/permission-grants.slice';
 import profile from '../slices/profile.slice';
 import proxyUrl from '../slices/proxy.slice';
 import resources from '../slices/resources.slice';
